+++ pkg/microservice/systemconfig/core/codehost/service/codehost.go
@@ -4,13 +4,13 @@ import (
 	"go.uber.org/zap"
 
 	"github.com/koderover/zadig/pkg/microservice/systemconfig/core/codehost/repository/models"
-	"github.com/koderover/zadig/pkg/microservice/systemconfig/core/codehost/repository/mongo"
+	"github.com/koderover/zadig/pkg/microservice/systemconfig/core/codehost/repository/mongodb"
 )
 
 func CreateCodehost(codehost *models.CodeHost, _ *zap.SugaredLogger) (*models.CodeHost, error) {
-	return mongo.NewCodehostColl().AddCodeHost(codehost)
+	return mongodb.NewCodehostColl().AddCodeHost(codehost)
 }
 
 func FindCodehost(_ *zap.SugaredLogger) ([]*models.CodeHost, error) {
-	return mongo.NewCodehostColl().FindCodeHosts()
+	return mongodb.NewCodehostColl().FindCodeHosts()
 }
