+++ pkg/microservice/systemconfig/core/email/repository/mongodb/emailservice.go
@@ -1,4 +1,4 @@
-package mongo
+package mongodb
 
 import (
 	"context"
