+++ pkg/microservice/systemconfig/core/jira/repository/mongodb/jira.go
@@ -1,4 +1,4 @@
-package mongo
+package mongodb
 
 import (
 	"context"
